LPS-136256 Rename object files
