+++ modules/apps/object/object-service/src/main/java/com/liferay/object/model/impl/ObjectDefinitionImpl.java
@@ -50,7 +50,7 @@ public class ObjectDefinitionImpl extends ObjectDefinitionBaseImpl {
 			throw new UnsupportedOperationException();
 		}
 
-		return ObjectPortletKeys.OBJECT + "#" + getObjectDefinitionId();
+		return ObjectPortletKeys.OBJECT_ENTRIES + "#" + getObjectDefinitionId();
 	}
 
 	@Override
