+++ modules/apps/object/object-service/src/main/resources/resource-actions/default.xml
@@ -5,7 +5,7 @@
 	<model-resource>
 		<model-name>com.liferay.object</model-name>
 		<portlet-ref>
-			<portlet-name>com_liferay_object_web_internal_portlet_ObjectAdminPortlet</portlet-name>
+			<portlet-name>com_liferay_object_web_internal_object_definitions_portlet_ObjectDefinitionsPortlet</portlet-name>
 		</portlet-ref>
 		<root>true</root>
 		<weight>1</weight>
@@ -23,7 +23,7 @@
 	<model-resource>
 		<model-name>com.liferay.object.model.ObjectDefinition</model-name>
 		<portlet-ref>
-			<portlet-name>com_liferay_object_web_internal_portlet_ObjectAdminPortlet</portlet-name>
+			<portlet-name>com_liferay_object_web_internal_object_definitions_portlet_ObjectDefinitionsPortlet</portlet-name>
 		</portlet-ref>
 		<weight>1</weight>
 		<permissions>
