--- modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/constants/ObjectAdminScreenNavigationEntryConstants.java
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.object.web.internal.constants;
-
-/**
- * @author Marco Leo
- * @author Gabriel Albuquerque
- */
-public class ObjectAdminScreenNavigationEntryConstants {
-
-	public static final String CATEGORY_KEY_DETAILS = "details";
-
-	public static final String CATEGORY_KEY_FIELDS = "fields";
-
-	public static final String ENTRY_KEY_DETAILS = "details";
-
-	public static final String ENTRY_KEY_FIELDS = "fields";
-
-	public static final String SCREEN_NAVIGATION_KEY_OBJECT_ADMIN =
-		"edit.object.admin.form";
-
-}
