+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/deployer/ObjectDefinitionDeployerImpl.java
@@ -21,9 +21,9 @@ import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchemaBuild
 import com.liferay.object.deployer.ObjectDefinitionDeployer;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.service.ObjectFieldLocalService;
-import com.liferay.object.web.internal.application.list.ObjectDefinitionPanelApp;
-import com.liferay.object.web.internal.frontend.taglib.clay.data.set.view.table.ObjectDefinitionTableClayDataSetDisplayView;
-import com.liferay.object.web.internal.portlet.ObjectPortlet;
+import com.liferay.object.web.internal.object.entries.application.list.ObjectEntriesPanelApp;
+import com.liferay.object.web.internal.object.entries.frontend.taglib.clay.data.set.view.table.ObjectEntriesTableClayDataSetDisplayView;
+import com.liferay.object.web.internal.object.entries.portlet.ObjectEntriesPortlet;
 import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
 import com.liferay.portal.kernel.util.Portal;
 
@@ -51,7 +51,7 @@ public class ObjectDefinitionDeployerImpl implements ObjectDefinitionDeployer {
 		return Arrays.asList(
 			_bundleContext.registerService(
 				ClayDataSetDisplayView.class,
-				new ObjectDefinitionTableClayDataSetDisplayView(
+				new ObjectEntriesTableClayDataSetDisplayView(
 					_clayTableSchemaBuilderFactory, objectDefinition,
 					_objectFieldLocalService.getObjectFields(
 						objectDefinition.getObjectDefinitionId())),
@@ -60,7 +60,7 @@ public class ObjectDefinitionDeployerImpl implements ObjectDefinitionDeployer {
 					objectDefinition.getPortletId()
 				).build()),
 			_bundleContext.registerService(
-				PanelApp.class, new ObjectDefinitionPanelApp(objectDefinition),
+				PanelApp.class, new ObjectEntriesPanelApp(objectDefinition),
 				HashMapDictionaryBuilder.<String, Object>put(
 					"panel.app.order:Integer", "300"
 				).put(
@@ -68,7 +68,7 @@ public class ObjectDefinitionDeployerImpl implements ObjectDefinitionDeployer {
 				).build()),
 			_bundleContext.registerService(
 				Portlet.class,
-				new ObjectPortlet(
+				new ObjectEntriesPortlet(
 					_portal, objectDefinition.getRESTContextPath()),
 				HashMapDictionaryBuilder.<String, Object>put(
 					"com.liferay.portlet.display-category", "category.hidden"
@@ -78,7 +78,8 @@ public class ObjectDefinitionDeployerImpl implements ObjectDefinitionDeployer {
 				).put(
 					"javax.portlet.name", objectDefinition.getPortletId()
 				).put(
-					"javax.portlet.init-param.view-template", "/object/view.jsp"
+					"javax.portlet.init-param.view-template",
+					"/object_entries/view_object_entries.jsp"
 				).build()));
 	}
