--- modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/frontend/taglib/servlet/taglib/BaseObjectAdminScreenNavigationEntry.java
@@ -1,57 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.object.web.internal.frontend.taglib.servlet.taglib;
-
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
-import com.liferay.object.model.ObjectDefinition;
-import com.liferay.object.web.internal.constants.ObjectAdminScreenNavigationEntryConstants;
-
-import java.io.IOException;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Marco Leo
- * @author Gabriel Albuquerque
- */
-public abstract class BaseObjectAdminScreenNavigationEntry
-	implements ScreenNavigationEntry<ObjectDefinition> {
-
-	public abstract String getJspPath();
-
-	@Override
-	public String getScreenNavigationKey() {
-		return ObjectAdminScreenNavigationEntryConstants.
-			SCREEN_NAVIGATION_KEY_OBJECT_ADMIN;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse, getJspPath());
-	}
-
-	@Reference
-	protected JSPRenderer jspRenderer;
-
-}
