--- modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/frontend/taglib/servlet/taglib/ObjectFieldAdminScreenNavigationCategory.java
@@ -1,81 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.object.web.internal.frontend.taglib.servlet.taglib;
-
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.object.web.internal.constants.ObjectAdminScreenNavigationEntryConstants;
-import com.liferay.object.web.internal.display.context.ObjectFieldAdminDisplayContext;
-import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
-
-import java.util.Locale;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Marco Leo
- * @author Gabriel Albuquerque
- */
-@Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
-)
-public class ObjectFieldAdminScreenNavigationCategory
-	extends BaseObjectAdminScreenNavigationEntry
-	implements ScreenNavigationCategory {
-
-	@Override
-	public String getCategoryKey() {
-		return ObjectAdminScreenNavigationEntryConstants.CATEGORY_KEY_FIELDS;
-	}
-
-	@Override
-	public String getEntryKey() {
-		return ObjectAdminScreenNavigationEntryConstants.ENTRY_KEY_FIELDS;
-	}
-
-	@Override
-	public String getJspPath() {
-		return "/object_admin/view_object_fields.jsp";
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "fields");
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT,
-			new ObjectFieldAdminDisplayContext(httpServletRequest));
-
-		super.render(httpServletRequest, httpServletResponse);
-	}
-
-}
