+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/object/definitions/application/list/ObjectDefinitionsPanelCategory.java
@@ -1,0 +1,62 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.object.web.internal.object.definitions.application.list;
+
+import com.liferay.application.list.BasePanelCategory;
+import com.liferay.application.list.PanelCategory;
+import com.liferay.application.list.constants.PanelCategoryKeys;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
+import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.service.permission.PortalPermissionUtil;
+
+import java.util.Locale;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Marco Leo
+ */
+@Component(
+	immediate = true,
+	property = {
+		"panel.category.key=" + PanelCategoryKeys.CONTROL_PANEL,
+		"panel.category.order:Integer=400"
+	},
+	service = PanelCategory.class
+)
+public class ObjectDefinitionsPanelCategory extends BasePanelCategory {
+
+	@Override
+	public String getKey() {
+		return PanelCategoryKeys.CONTROL_PANEL_OBJECT;
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		return LanguageUtil.get(locale, "object");
+	}
+
+	@Override
+	public boolean isShow(PermissionChecker permissionChecker, Group group)
+		throws PortalException {
+
+		return PortalPermissionUtil.contains(
+			permissionChecker, ActionKeys.VIEW_CONTROL_PANEL);
+	}
+
+}
