--- modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/portlet/action/EditObjectAdminMVCRenderCommand.java
@@ -1,69 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.object.web.internal.portlet.action;
-
-import com.liferay.object.constants.ObjectPortletKeys;
-import com.liferay.object.service.ObjectDefinitionLocalService;
-import com.liferay.object.web.internal.constants.ObjectWebKeys;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
-import com.liferay.portal.kernel.servlet.SessionErrors;
-import com.liferay.portal.kernel.util.ParamUtil;
-
-import javax.portlet.PortletException;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Marco Leo
- * @author Gabriel Albuquerque
- */
-@Component(
-	property = {
-		"javax.portlet.name=" + ObjectPortletKeys.OBJECT_ADMIN,
-		"mvc.command.name=/object_admin/edit_object_admin"
-	},
-	service = MVCRenderCommand.class
-)
-public class EditObjectAdminMVCRenderCommand implements MVCRenderCommand {
-
-	@Override
-	public String render(
-			RenderRequest renderRequest, RenderResponse renderResponse)
-		throws PortletException {
-
-		long objectDefinitionId = ParamUtil.getLong(
-			renderRequest, "objectDefinitionId");
-
-		try {
-			renderRequest.setAttribute(
-				ObjectWebKeys.OBJECT_DEFINITION,
-				_objectDefinitionLocalService.getObjectDefinition(
-					objectDefinitionId));
-		}
-		catch (PortalException portalException) {
-			SessionErrors.add(renderRequest, portalException.getClass());
-		}
-
-		return "/object_admin/edit_object.jsp";
-	}
-
-	@Reference
-	private ObjectDefinitionLocalService _objectDefinitionLocalService;
-
-}
