--- modules/apps/object/object-web/src/main/resources/META-INF/resources/object/view.jsp
@@ -1,33 +1,0 @@
-<%--
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
---%>
-
-<%@ include file="/init.jsp" %>
-
-<%
-ObjectDisplayContext objectDisplayContext = (ObjectDisplayContext)request.getAttribute(WebKeys.PORTLET_DISPLAY_CONTEXT);
-%>
-
-<clay:headless-data-set-display
-	apiURL="<%= objectDisplayContext.getAPIURL() %>"
-	clayDataSetActionDropdownItems="<%= objectDisplayContext.getClayDataSetActionDropdownItems() %>"
-	formId="fm"
-	id="<%= objectDisplayContext.getClayHeadlessDataSetDisplayId() %>"
-	itemsPerPage="<%= 20 %>"
-	namespace="<%= liferayPortletResponse.getNamespace() %>"
-	pageNumber="<%= 1 %>"
-	portletURL="<%= liferayPortletResponse.createRenderURL() %>"
-	style="fluid"
-/>
