--- modules/apps/object/object-web/src/main/resources/META-INF/resources/object_admin/edit_object.jsp
@@ -1,35 +1,0 @@
-<%--
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
---%>
-
-<%@ include file="/init.jsp" %>
-
-<%
-ObjectDefinition objectDefinition = (ObjectDefinition)request.getAttribute(ObjectWebKeys.OBJECT_DEFINITION);
-%>
-
-<liferay-frontend:screen-navigation
-	context="<%= objectDefinition %>"
-	key="<%= ObjectAdminScreenNavigationEntryConstants.SCREEN_NAVIGATION_KEY_OBJECT_ADMIN %>"
-	portletURL='<%=
-		PortletURLBuilder.createRenderURL(
-			renderResponse
-		).setMVCRenderCommandName(
-			"/object_admin/edit_object_admin"
-		).setParameter(
-			"objectDefinitionId", objectDefinition.getObjectDefinitionId()
-		).build()
-	%>'
-/>
