+++ modules/apps/object/object-web/src/main/resources/content/Language.properties
@@ -1,7 +1,7 @@
 add-object=Add Object
 create-new-object=Create New Object
-javax.portlet.title.com_liferay_object_web_internal_portlet_ObjectAdminPortlet=Objects
-javax.portlet.title.com_liferay_object_web_internal_portlet_ObjectPortlet=Objects
+javax.portlet.title.com_liferay_object_web_internal_object_definitions_portlet_ObjectDefinitionsPortlet=Objects
+javax.portlet.title.com_liferay_object_web_internal_object_entries_portlet_ObjectEntriesPortlet=Objects
 model.resource.com.liferay.object.model.ObjectEntry=Object
 object=Object
 object-definition-data=Object Data
